+++ integration-tests/worker/test/integration.test.ts
@@ -313,7 +313,7 @@ test.serial('run a job with bad credentials', (t) => {
       t.is(payload.error_type, 'CredentialLoadError');
       t.regex(
         payload.error_message,
-        /Failed to load credential zzz: not_found/
+        /Failed to load credential zzz: \[fetch:credential\] not_found/
       );
       done();
     });
@@ -554,6 +554,31 @@ test.serial("Don't send job logs to stdout", (t) => {
         (l) => l.name === 'engine' && l.message[0].match(/complete workflow/i)
       );
       t.truthy(runtimeLog);
+      done();
+    });
+
+    lightning.enqueueRun(attempt);
+  });
+});
+
+// This is a test against job logs - but it should work
+// exactly the same way for all logs
+test.serial("Don't send empty logs to lightning", (t) => {
+  return new Promise(async (done) => {
+    const attempt = {
+      id: crypto.randomUUID(),
+      jobs: [
+        {
+          adaptor: '@openfn/language-common@latest',
+          body: 'fn((s) =>  { console.log(); return s })',
+        },
+      ],
+    };
+
+    lightning.once('run:complete', () => {
+      // The engine logger shouldn't print out any job logs
+      const jobLogs = engineLogger._history.filter((l) => l.name === 'JOB');
+      t.is(jobLogs.length, 0);
       done();
     });